BUILD: strip preview on-the-fly
